+++ platform/mv3/extension/js/scripting-manager.js
@@ -519,3 +519,8 @@ export async function onWakeupRun() {
 }
 
 /******************************************************************************/
+
+export {
+    registerInjectables,
+		registerAdNauseam // ADN
+};